Updated UI, added settings
