--- Assets/Scenes/BScenes/Prefab/Ingame_UI.prefab.meta
@@ -1,7 +1,0 @@
-fileFormatVersion: 2
-guid: e062e1bf044171748b4c65294d6f3c2c
-PrefabImporter:
-  externalObjects: {}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
