+++ Assets/Scenes/BScenes/Prefab/Managers.prefab.meta
@@ -1,0 +1,7 @@
+fileFormatVersion: 2
+guid: 5dba6b50f32bdf940978f458b99d300a
+PrefabImporter:
+  externalObjects: {}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
