--- Assets/Scenes/BScenes/Prefab/Pause.prefab.meta
@@ -1,7 +1,0 @@
-fileFormatVersion: 2
-guid: e212fc468f2dcc842a5d16fae0ffd5a4
-PrefabImporter:
-  externalObjects: {}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
