+++ Assets/Scenes/BScenes/Prefab/UI.meta
@@ -1,0 +1,8 @@
+fileFormatVersion: 2
+guid: 50145560ae5bda340a7fbffae0aaccf9
+folderAsset: yes
+DefaultImporter:
+  externalObjects: {}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
