+++ Assets/Scenes/BScenes/Prefab/UI/Ingame_UI.prefab.meta
@@ -1,0 +1,7 @@
+fileFormatVersion: 2
+guid: e062e1bf044171748b4c65294d6f3c2c
+PrefabImporter:
+  externalObjects: {}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
