+++ Assets/Scenes/BScenes/Prefab/UI/Pause_Panel.prefab.meta
@@ -1,0 +1,7 @@
+fileFormatVersion: 2
+guid: f2bff2ce5aa92684cbb9eec20cc196d7
+PrefabImporter:
+  externalObjects: {}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
