--- Assets/Scenes/BScenes/Scripts/DamageHandler.cs.meta
@@ -1,11 +1,0 @@
-fileFormatVersion: 2
-guid: 74fcce5a535cf794d82106c4d6b17b96
-MonoImporter:
-  externalObjects: {}
-  serializedVersion: 2
-  defaultReferences: []
-  executionOrder: 0
-  icon: {instanceID: 0}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
