+++ Assets/Scenes/BScenes/Scripts/Handler.meta
@@ -1,0 +1,8 @@
+fileFormatVersion: 2
+guid: 66f126a53785a7c43b9d2ae9d13f935d
+folderAsset: yes
+DefaultImporter:
+  externalObjects: {}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
