--- Assets/Scenes/BScenes/Scripts/SkillTree.cs.meta
@@ -1,11 +1,0 @@
-fileFormatVersion: 2
-guid: 23af47e764a28644487f7607890bd378
-MonoImporter:
-  externalObjects: {}
-  serializedVersion: 2
-  defaultReferences: []
-  executionOrder: 0
-  icon: {instanceID: 0}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
