+++ Assets/Scenes/BScenes/Scripts/UI/Controls.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: c503160de5d5dfe468187e5dab170891
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
