+++ Assets/Scenes/BScenes/Scripts/UI/SkillButton.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: d93c1b20c553f5845ac10f8075ba0fd7
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
