+++ Assets/Scenes/BScenes/Scripts/UI/SkillTree.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: 4b7c20a28727f2b469dcd3530ac511c6
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
