--- Assets/SkillButton.cs.meta
@@ -1,11 +1,0 @@
-fileFormatVersion: 2
-guid: fc578fbd130d2cb41bd716b582171b6b
-MonoImporter:
-  externalObjects: {}
-  serializedVersion: 2
-  defaultReferences: []
-  executionOrder: 0
-  icon: {instanceID: 0}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
