+++ ProjectSettings/TagManager.asset
@@ -7,7 +7,6 @@ TagManager:
   - Ground
   - Enemy
   - Card
-  - dangerous
   - Door
   - Chip_01
   - Chip_02
